Split benchmark suite for select and slice operators
